Fix USERID in build_docker if script is called as superuser (#55)
* fixes readme.md link to Dockerfile

Dockerfile is spelled with capital D in filepath, but not in readme.md
This commit fixes the link in the readme.md file

* fixes USERID lookup in build_docker.sh if docker daemon requires sudo

In case the docker daemon requires sudo privileges to run, calling "sudo
build_docker.sh" breaks the user id lookup. This commit adds a check to
see if the buildscript is run as su, and updates USERID to the original
callers id.

---------

Co-authored-by: pHellr <>
Co-authored-by: pHellr <pHellr@users.noreply.github.com>